refactored type declarations, moved out dependencies on ract (for expedition service model)
