+++ app/actions/Web.tsx
@@ -6,7 +6,8 @@ import {initQuest} from './Quest'
 import {openSnackbar} from '../actions/Snackbar'
 import {userFeedbackClear} from '../actions/UserFeedback'
 import {SearchSettings, SettingsType, QuestState, UserState, UserFeedbackState} from '../reducers/StateTypes'
-import {QuestContext, QuestDetails, defaultQuestContext} from '../reducers/QuestTypes'
+import {QuestContext, QuestDetails} from '../reducers/QuestTypes'
+import {defaultQuestContext} from '../reducers/Quest'
 import {getDevicePlatform, getAppVersion} from '../Globals'
 import {logEvent} from '../React'
