+++ app/cardtemplates/combat/Actions.tsx
@@ -1,6 +1,7 @@
 import Redux from 'redux'
-import {DifficultyType, defaultQuestContext, Enemy, Loot} from '../../reducers/QuestTypes'
-import {CombatDifficultySettings, CombatAttack} from './State'
+import {DifficultyType, Enemy, Loot} from '../../reducers/QuestTypes'
+import {defaultQuestContext} from '../../reducers/Quest'
+import {CombatDifficultySettings, CombatAttack} from './Types'
 import {SettingsType} from '../../reducers/StateTypes'
 import {ParserNode} from '../../parser/Node'
 import {toCard} from '../../actions/Card'
